refactor: Implement rewritten frontend
