+++ components/nebulaStats.js
@@ -5,14 +5,12 @@ import styled from "styled-components";
 
 const Wrapper = styled.section`
   position: relative;
-  margin-top: 70px;
-  margin-left: 14px;
-`;
-
-const Div = styled.div`
-  margin-left: 1px;
   display: flex;
   flex-direction: column;
+`;
+
+const StyledIon = styled(Ion)`
+  margin-top: 15.5px;
 `;
 
 function NebulaStats() {
@@ -21,10 +19,7 @@ function NebulaStats() {
       <Wrapper>
         <h1>NebulaStats</h1>
         <Stats />
-        <Div>
-        <Ion ionName="John's Phone" ionStatus={true} />
-        </Div>
-        
+        <StyledIon ionName="John's Phone" ionStatus={true} />
       </Wrapper>
     </>
   );
